requirements.txt
PyYAML
rdflib